Regression Test for Count

Perform Poisson and Negative Binomial regression analysis to compare the counts from different groups, treatment and control. The difference between functions 'countTest2' and 'countTest' resides in the estimation of the prior weights used in Negative Binomial generalized linear model.

countTest2(
  DS,
  num.cores = 1,
  countFilter = TRUE,
  CountPerBp = NULL,
  minCountPerIndv = 3,
  maxGrpCV = NULL,
  FilterLog2FC = TRUE,
  pAdjustMethod = "BH",
  pvalCutOff = 0.05,
  MVrate = 0.98,
  Minlog2FC = 0.5,
  test = c("Wald", "LRT"),
  scaling = 1L,
  tasks = 0L,
  saveAll = FALSE,
  verbose = TRUE
)

Arguments

DS
A 'glmDataSet' object, which is created with function glmDataSet.

num.cores, tasks
Parameters for parallel computation using package BiocParallel-package: the number of cores to use, i.e. at most how many child processes will be run simultaneously (see bplapply and the number of tasks per job (only for Linux OS).

countFilter
whether or not to filter the counts according to the minimum count per region per each individual/sample, which is setting by 'minCountPerIndv'.

CountPerBp
for each group the count per bp must be equal or greater than CountPerBp. The filter is applied if 'CountPerBp' is given and if 'x' DESeqDataSet object has the rowRanges as a GRanges object on it

minCountPerIndv
each gene or region must have more than 'minCountPerIndv' counts (on average) per individual in at least one group.

maxGrpCV
A numerical vector. Maximum coefficient of variance for each group. Default maxGrpCV = NULL. The numbers 'maxGrpCV1' and 'maxGrpCV2' will be taken as the maximun variances values permitted in control and in treatment groups, respectively. If only 'maxGrpCV1' is provided, then maxGrpCV = c('maxGrpCV1', 'maxGrpCV1'). This parameter is addressed to prevent testing regions where intra-group variations are very large, e.g.: control = c(1,0,1,1) and treatment = c(1, 0, 1, 40). The coefficient of variance for the treatment group is 1.87, very high. The generalized linear regression analysis would yield statistical significant group differences, but evidently there is something wrong in one of the treatment samples. We would try the application of further statistical smoothing approach, but we prefer to leave the user decide which regions to test.

FilterLog2FC
if TRUE, the results are filtered using the minimum absolute value of log2FoldChanges observed to accept that a gene in the treatment is differentially expressed in respect to the control

pAdjustMethod
method used to adjust the results; default: BH

pvalCutOff
cutoff used then a p-value adjustment is performed

MVrate
Minimum Mean/Variance rate.

Minlog2FC
minimum logarithm base 2 of fold changes.

test
A character string matching one of 'Wald' or 'LRT'. If test = 'Wald', then the p-value of the Wald test for the coefficient of the independent variable (treatment group) will be reported. If test = 'LRT', then the p-value from a likelihood ratio test given by anova function from stats packages will be the reported p-value for the group comparison when the best fitted model is the negative binomial. As suggested for glm, if best fitted model is Poisson or quasi-Poisson, then the best test is 'Chi-squared' or 'F-test', respectively. So, for the sake of simplicity, the corresponding suitable test will be applied when test = 'LRT'.

scaling
integer (default 1). Scaling factor estimate the signal density as: scaling x 'DMP-Count-Per-Bp'. For example, if scaling = 1000, then signal density denotes the number of DMPs in 1000 bp.

saveAll
if TRUE all the temporal results are returned

verbose
if TRUE, prints the function log to stdout

Value

a data frame or GRanges object (if the DS contain the GRanges information for each gene) with the test results and original count matrix, plus control and treatment signal densities and their variation.

Details

A pairwise group comparison, control versus treatment, is performed. The experimental design settings must be introduced using function glmDataSet to provide dataset (DS) object.

See also

Examples

set.seed(133) # Set a seed
## A GRanges object with the count matrix in the metacolumns is created
countData <- matrix(sample.int(200, 500, replace = TRUE), ncol = 4)
colnames(countData) <- c('A1','A2','B1','B2')

start <- seq(1, 25e4, 2000)
end <- start + 1000
chr <- c(rep('chr1', 70), rep('chr2', 55))
GR <- GRanges(seqnames = chr, IRanges(start = start, end = end))
mcols(GR) <- countData

## Gene IDs
names(GR) <- paste0('gene', 1:length(GR))

## An experiment design is set.
colData <- data.frame(condition = factor(c('A','A','B','B')),
                    c('A1','A2','B1','B2'), row.names = 2)

## A RangedGlmDataSet is created
ds <- glmDataSet(GR = GR, colData = colData)

## The gneralized linear model pairwise group comparison, group 'A'
## ('control') versus 'B' (treatment) is performed.
countTest2(ds, num.cores = 1L, maxGrpCV = c(0.4, 0.4), verbose = FALSE)

#> GRanges object with 10 ranges and 12 metadata columns:
#>          seqnames        ranges strand |        A1        A2        B1
#>             <Rle>     <IRanges>  <Rle> | <integer> <integer> <integer>
#>    gene5     chr1     8001-9001      * |        43        42        72
#>   gene36     chr1   70001-71001      * |       165       187        25
#>   gene38     chr1   74001-75001      * |       117       165        27
#>   gene41     chr1   80001-81001      * |        76       102       198
#>   gene42     chr1   82001-83001      * |       142       188        86
#>   gene47     chr1   92001-93001      * |        24        16       101
#>   gene69     chr1 136001-137001      * |        49        75       110
#>   gene72     chr2 142001-143001      * |        51        38       112
#>   gene75     chr2 148001-149001      * |       195       190        26
#>   gene78     chr2 154001-155001      * |       117        80       184
#>                 B2             log2FC  scaled.deviance               pvalue
#>          <integer>          <numeric>        <numeric>            <numeric>
#>    gene5       128   0.85566611005772 14.0356701793316 0.000180245933077019
#>   gene36        35    -1.769286613376 25.1439698545735 5.36539225149968e-06
#>   gene38        17  -1.85771743701985 6.79422202295607    0.013376785333257
#>   gene41       126  0.598959965500245 9.38158872001879   0.0021460958381637
#>   gene42       113 -0.505787829736033 13.1528553754701 0.000295434175502001
#>   gene47        72   1.46441214038384 7.91261025502834  0.00689645785092455
#>   gene69       129  0.656181986326474 22.5625506376617 2.62351534194847e-06
#>   gene72        96  0.848901709969179 9.92488082051744  0.00219603135316477
#>   gene75        40  -1.76358859226136 19.7071745248823 5.06968870813117e-05
#>   gene78       186  0.630299276900281 5.19588591833575   0.0236384622748537
#>                   model             adj.pval   CT.SignalDensity
#>                <factor>            <numeric>          <numeric>
#>    gene5   Neg.Binomial 0.000450614832692548 0.0424575424575425
#>   gene36 Neg.Binomial.W 2.68269612574984e-05  0.175824175824176
#>   gene38 Neg.Binomial.W     0.01486309481473  0.140859140859141
#>   gene41   Neg.Binomial  0.00313718764737824 0.0889110889110889
#>   gene42   Neg.Binomial 0.000590868351004001  0.164835164835165
#>   gene47 Neg.Binomial.W  0.00862057231365569   0.01998001998002
#>   gene69   Neg.Binomial 2.62351534194847e-05 0.0619380619380619
#>   gene72 Neg.Binomial.W  0.00313718764737824 0.0444555444555445
#>   gene75 Neg.Binomial.W 0.000168989623604372  0.192307692307692
#>   gene78 Neg.Binomial.W   0.0236384622748537 0.0984015984015984
#>            TT.SignalDensity SignalDensityVariation
#>                   <numeric>              <numeric>
#>    gene5 0.0999000999000999     0.0574425574425574
#>   gene36   0.02997002997003     -0.145854145854146
#>   gene38  0.021978021978022     -0.118881118881119
#>   gene41  0.161838161838162     0.0729270729270729
#>   gene42 0.0994005994005994    -0.0654345654345654
#>   gene47 0.0864135864135864     0.0664335664335664
#>   gene69  0.119380619380619     0.0574425574425574
#>   gene72  0.103896103896104     0.0594405594405594
#>   gene75  0.032967032967033     -0.159340659340659
#>   gene78  0.184815184815185     0.0864135864135864
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths