DMP clustering

Given a 'pDMP' (or 'InfDiv') object carrying DMPs (methylated cytosines) detected in Methyl-IT downstream analysis, function 'dmpClusters' build clusters of DMPs, which can be further tested to identify differentially methylated regions (DMRs) with countTest2 function.

dmpClusters(
  GR,
  maxDist = 3,
  minNumDMPs = 1,
  maxClustDist = NULL,
  method = c("relaxed", "fixed.int"),
  chromosomes = NULL,
  columns = 9L,
  ignore.strand = TRUE,
  num.cores = 1L,
  tasks = 0L,
  verbose = TRUE,
  ...
)

Arguments

GR
An object from 'pDMP' class, which is returned by selectDIMP function.

maxDist
maximum distance at which two reported bases sites from the same cluster can be separated. Default: \(maxDist = 3\).

minNumDMPs
Minimum number of DMPs inside of each cluster. Default: \(minNumDMPs = 1\).

maxClustDist
Clusters separated by a distance lesser than 'maxClustDist' positions are merged. Default: \(maxClustDist = NULL\). If \(0 < maxClustDist < maxDist\), then maxClustDist will be recalculated as \(maxClustDist = maxDist + 1\).

method
Two different approaches are implemented to clustering DMPs:

"relaxed":

DMP ranges which are separated by a distance less than 'maxClustDist' are merged and ranges with less than 'minNumDMPs' are removed.

"fixed.int":

It will generate a partition where the distance between consecutive DMPs is not greater than 'maxDist'. Ranges with less than 'minNumDMPs' are removed. If, additionally, a value maxClustDist > 0 is provided, then the "relaxed" approach is applied to the ranges from the first step.

chromosomes
vector of characters labeling the chromosomes included in the analysis. Default: chromosomes = NULL (all chromosomes are included).

columns
An integer number corresponding to the specific column(s) to use from the meta-column of each GRanges. Default values is 9 (the column carrying to the Hellinger divergence values).

ignore.strand
Same as in findOverlaps-methods.

num.cores, tasks
integer(1). The number of cores to use, i.e. at most how many child processes will be run simultaneously (see bplapply function from BiocParallel package).The number of tasks per job. value must be a scalar integer >= 0L (see MulticoreParam from BiocParallel package).

verbose
if TRUE, prints the function log to stdout.

...
Further parameters for uniqueGRanges function.

Value

A GRanges object with the numbers of positions inside each cluster, where DMPs were reported in at least one of the samples.

Details

Two algorithmic approaches are implemented, named: "relaxed" and "fixed.int" (see the description of parameter 'method'). The "fixed.int" is mostly addressed to find specific methylation patterns, but the price is the number of DMRs found is lower.

The number of DMPs reported in each cluster corresponds to the numbers of sites inside the cluster where DMPs were found in at least one of the samples (from control or from treatment). That is, dmpClusters is only a tool to locate regions with high density of DMPs from all the samples. It does not detect DMRs. It is assumed that only DMP coordinates are given in the 'GR' object. That is, all the sites provided are considered in the computation.

References

Sanchez R, Mackenzie SA (2016) Information Thermodynamics of Cytosine DNA Methylation. PLOS ONE 11(3): e0150427. https://doi.org/10.1371/journal.pone.0150427

Examples

## Get a dataset of dmps from the package
data(dmps)

## Build clusters of DMPs taking into account the DNA strand
x1 = dmpClusters2(GR = dmps, maxDist = 7, minNumDMPs = 6,
                 method = "fixed.int", ignore.strand = FALSE,
                 verbose = FALSE)

#> Error in dmpClusters2(GR = dmps, maxDist = 7, minNumDMPs = 6, method = "fixed.int",     ignore.strand = FALSE, verbose = FALSE): could not find function "dmpClusters2"

data.frame(x1)

#> Error in data.frame(x1): object 'x1' not found

# \donttest{
## Build clusters of DMPs ignoring DNA strand and maxClustDist = 7
x2 = dmpClusters2(GR = dmps, maxDist = 7, minNumDMPs = 6,
                  maxClustDist = 7, method = "fixed.int",
                  num.cores=2L, ignore.strand = TRUE,
                  verbose = FALSE)

#> Error in dmpClusters2(GR = dmps, maxDist = 7, minNumDMPs = 6, maxClustDist = 7,     method = "fixed.int", num.cores = 2L, ignore.strand = TRUE,     verbose = FALSE): could not find function "dmpClusters2"

DataFrame(data.frame(x2))

#> Error in data.frame(x2): object 'x2' not found

## The relaxed approach with method = "relaxed"
x3 = dmpClusters2(GR = dmps, minNumDMPs = 6, method = "relaxed",
                  maxClustDist = 10, ignore.strand = TRUE,
                  verbose = FALSE)

#> Error in dmpClusters2(GR = dmps, minNumDMPs = 6, method = "relaxed", maxClustDist = 10,     ignore.strand = TRUE, verbose = FALSE): could not find function "dmpClusters2"

DataFrame(data.frame(x3))

#> Error in data.frame(x3): object 'x3' not found

## ==== Setting up the experiment design to test for DMRs ===
nams <- names(dmps)
dmps_at_clusters <- getDMPatRegions(GR = dmps, regions = x3,
                                    ignore.strand = TRUE)

#> Error in FUN(X[[i]], ...): object 'x3' not found

dmps_at_clusters <- uniqueGRanges(dmps_at_clusters, columns = 2L,
                                ignore.strand = TRUE, type = 'equal',
                                verbose = FALSE)

#> Error in is(ListOfGranges, "list"): object 'dmps_at_clusters' not found

colnames(mcols(dmps_at_clusters)) <- nams

#> Error in colnames(mcols(dmps_at_clusters)) <- nams: object 'dmps_at_clusters' not found

colData <- data.frame(condition = factor(c('CT', 'CT', 'CT',
                                        'TT', 'TT', 'TT'),
                                        levels = c('CT', 'TT')),
                    nams, row.names = 2)

## Build a RangedGlmDataSet object
ds <- glmDataSet(GR = dmps_at_clusters, colData = colData)

#> Error in glmDataSet(GR = dmps_at_clusters, colData = colData): object 'dmps_at_clusters' not found

## ================ Testing to detect DMRs ===========
dmrs <- countTest2(DS = ds, num.cores = 4L, minCountPerIndv = 4,
                maxGrpCV = c(1, 1), Minlog2FC = 0.5,
                CountPerBp = 0.001, test = 'LRT',
                verbose = TRUE)

#> *** Number of GR after filtering counts 92

#> *** GLM...

dmrs

#> GRanges object with 17 ranges and 12 metadata columns:
#>           seqnames        ranges strand |        A1        A2        B1
#>              <Rle>     <IRanges>  <Rle> | <integer> <integer> <integer>
#>    gene22     chr1   42001-43001      * |        44        53       187
#>    gene26     chr1   50001-51001      * |       174       139        32
#>    gene29     chr1   56001-57001      * |        12        19       129
#>    gene57     chr1 112001-113001      * |        25        94       187
#>    gene58     chr1 114001-115001      * |       131        70        28
#>       ...      ...           ...    ... .       ...       ...       ...
#>   gene101     chr2 200001-201001      * |       166       168       117
#>   gene104     chr2 206001-207001      * |       116        64       152
#>   gene106     chr2 210001-211001      * |       122        85         3
#>   gene107     chr2 212001-213001      * |       181       111        76
#>   gene114     chr2 226001-227001      * |        33        34       111
#>                  B2    log2FC scaled.deviance      pvalue        model
#>           <integer> <numeric>       <numeric>   <numeric>  <character>
#>    gene22        85   1.03109        11.57051 6.70061e-04 Neg.Binomial
#>    gene26        62  -1.20291        30.96629 2.62550e-08 Neg.Binomial
#>    gene29        64   1.82870        29.49876 5.59525e-08 Neg.Binomial
#>    gene57       174   1.10975         6.43212 1.12075e-02 Neg.Binomial
#>    gene58        41  -1.06920        15.04312 1.05082e-04 Neg.Binomial
#>       ...       ...       ...             ...         ...          ...
#>   gene101        71 -0.574699        13.08305 2.97980e-04 Neg.Binomial
#>   gene104       146  0.504137         6.60778 1.01534e-02 Neg.Binomial
#>   gene106        16 -2.388280        30.98220 2.60406e-08 Neg.Binomial
#>   gene107        21 -1.102043         5.75603 1.64322e-02 Neg.Binomial
#>   gene114       158  1.390019        64.70832 8.68475e-16 Neg.Binomial
#>              adj.pval CT.SignalDensity TT.SignalDensity SignalDensityVariation
#>             <numeric>        <numeric>        <numeric>              <numeric>
#>    gene22 2.37099e-03        0.0484515        0.1358641              0.0874126
#>    gene26 2.41546e-07        0.1563437        0.0469530             -0.1093906
#>    gene29 4.28969e-07        0.0154845        0.0964036              0.0809191
#>    gene57 3.22215e-02        0.0594406        0.1803197              0.1208791
#>    gene58 4.83379e-04        0.1003996        0.0344655             -0.0659341
#>       ...         ...              ...              ...                    ...
#>   gene101 1.14226e-03        0.1668332       0.09390609             -0.0729271
#>   gene104 3.11371e-02        0.0899101       0.14885115              0.0589411
#>   gene106 2.41546e-07        0.1033966       0.00949051             -0.0939061
#>   gene107 4.44635e-02        0.1458541       0.04845155             -0.0974026
#>   gene114 1.33166e-14        0.0334665       0.13436563              0.1008991
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths

# }