Evaluate a model using the Akaike information criterion (AIC)

Evaluate a glm object using the Akaike information criterion (AIC)

.evaluateModel(model, test = c("Wald", "LRT"), p.value = 1)

Arguments

model

glm object

test

A character string matching one of 'Wald' or 'LRT'. If test = 'Wald', then the p-value of the Wald test for the coefficient of the independent variable (treatment group) will be reported. If test = 'LRT', then the p-value from a likelihood ratio test given by anova function from stats packages will be the reported p-value for the group comparison when the best fitted model is the negative binomial. As suggested for glm, if best fitted model is Poisson or quasi-Poisson, then the best test is 'Chi-squared' or 'F-test', respectively. So, for the sake of simplicity, the corresponding suitable test will be applied when test = 'LRT'.

p.value

Cut off p-value to reject the null hypothesis

Value

AIC value