DMGs Estimation on Specified Genomic Region

This function is a wrapping of several functions called in through the steps to estimate differentially methylated regions (DMGs).

getDMGs(dmps, regions, ...)

# S4 method for pDMP,GRanges
getDMGs(
  dmps,
  regions,
  chr = NULL,
  colData,
  gene_id_col = NULL,
  gene_name_col = NULL,
  by.coord = FALSE,
  ignore.strand = TRUE,
  type = "within",
  only.hypo = FALSE,
  only.hyper = FALSE,
  num.cores = 1L,
  countFilter = TRUE,
  CountPerBp = NULL,
  minCountPerIndv = 3,
  maxGrpCV = NULL,
  FilterLog2FC = TRUE,
  pAdjustMethod = "BH",
  pvalCutOff = 0.05,
  MVrate = 0.98,
  Minlog2FC = 0.5,
  test = c("Wald", "LRT"),
  scaling = 1L,
  tasks = 0L,
  saveAll = FALSE,
  verbose = TRUE
)

Arguments

dmps

An object from pDMP or InfDiv class, which are generated by functions: getPotentialDIMP and estimateDivergence, respectively.

regions

A GRanges object with gene coordinates and gene IDs. A column named 'gene_id' carrying the gene ids should be included in the metacolumns. If the meta-column named 'gene_id' is not provided, then gene (region) ids will be created using the gene (region) coordinates.

chr

Optional. A character string denoting the chromosome from the regions (i.e., the seqnames) where the DMGs will be estimated.

colData

The same as in glmDataSet. A data frame with one column named 'condition', which must be a factor with exactly two levels. The rownames of colData individual samples. The row names of colData must correspond to the column names of the count matrix.

gene_id_col

Optional. An integer denoting the column from the GENES metacolumn where the gene ids are given.

gene_name_col

Optional. An integer denoting the column from the GENES metacolumn where the gene 'name' are given.

by.coord

logical(1). If TRUE, then the DMP are count per coordinate and not per gene id.

ignore.strand, type

Same as for findOverlaps-methods.

only.hypo, only.hyper

logical(1). Whether to select only hypo-methylated or hyper-methylated cytosine sites.

num.cores, countFilter, CountPerBp, minCountPerIndv, maxGrpCV

The same as in function countTest2.

FilterLog2FC, pAdjustMethod, pvalCutOff, MVrate, Minlog2FC

The same as in function countTest2.

test, scaling, tasks, saveAll, verbose

The same as in function countTest2.

See also