Count DMPs at Genomic Regions

The function counts DMPs overlapping with genomic regions. In fact, this function operates as getDIMPatGenes function, but without the restrictions set for GRanges objects derived from MethylIT pipeline.

getDMPatRegions(GR, regions, type = "within", ignore.strand = TRUE, ...)

# S3 method for default
getDMPatRegions(GR, regions, type = "within", ignore.strand = TRUE, ...)

# S3 method for GRanges
getDMPatRegions(GR, regions, type = "within", ignore.strand = TRUE, ...)

# S3 method for pDMP
getDMPatRegions(GR, regions, type = "within", ignore.strand = TRUE, ...)

# S3 method for InfDiv
getDMPatRegions(GR, regions, type = "within", ignore.strand = TRUE, ...)

# S3 method for list
getDMPatRegions(GR, regions, type = "within", ignore.strand = TRUE, ...)

Arguments

GR
An objects object from the any of the classes: 'pDMP', 'InfDiv', GRangesList, GRanges or a list of GRanges.

regions
A GRanges object with gene coordinates and gene IDs. A meta-column named 'gene_id' carying the gene ids should be included. If the meta-column named 'gene_id' is not provided, then gene (region) ids will be created using the gene (region) coordinates.

ignore.strand, type
Same as for findOverlaps-methods

...
optional arguments for findOverlaps-methods. Users must evaluate whether specific setting makes sense on each particular context.

Value

A GRanges object

See also

Examples

## Gene annotation
genes <- GRanges(seqnames = '1',
                ranges = IRanges(start = c(3631, 6788, 11649),
                end = c(5899, 9130, 13714)),
                strand = c('+', '-', '-'))
mcols(genes) <- data.frame(gene_id = c('AT1G01010', 'AT1G01020',
                                        'AT1G01030'))

## Get a dataset of dmps from the package
data(dmps)

## Finally DMPs found on genes
dmrs <- getDMPatRegions(GR = dmps, regions = genes)