Get Supplemental Files from GEO

Decompress 'gzip' files.

getGEOSuppFiles(
  GEO,
  makeDirectory = FALSE,
  baseDir = getwd(),
  pattern = NULL,
  verbose = TRUE
)

Arguments

GEO

A character vector with GEO accession numbers.

makeDirectory

Logic (FALSE). If GEO accession number is provided, whether to create a subdirectory for the downloaded files.

baseDir

Directory where files are downloads if GEO accession number is provided. Default is the current working directory.

pattern

A pattern for the name of the supplementary files from the GEO dataset. If provided, then only the files with the given pattern are downloaded. Otherwise, all the supplementary files are downloaded.

verbose

If TRUE, prints the function log to stdout

Value

A data frame is returned invisibly with rownames representing the full path of the resulting downloaded files and the records in the data.frame the output of file.info for each downloaded file.

Details

Download supplemental files from a specified GEO dataset. This function is originally provided in the Bioconductor package 'GEOquery'. The original function download all the supplemental files for a given GEO accession number. Herein small detail is added to permit only the download of the specified files and from several GEO accession numbers with only one call to the function.

Author

Original author: Sean Davis sdavis2@mail.nih.gov

Examples

## Download supplementary files from GEO data set and store fullpath/name
## in variable filename. The parameter 'pattern' permits us to download only
## the specified filesCG, in this case, CG and CHG methylation contexts.

if (FALSE) {
 filenames <- getGEOSuppFiles(GEO = 'GSM881757',
                pattern = 'G_cytosine.txt.gz')

 file.remove(filenames) ## Remove the downloaded file
}