Helmholtz Free Energy of Generalized Gamma Distribution

This function computes the Helmholtz Free Energy (HFE) for member of the Generalized Gamma (GG) Distribution family. The Helmholtz free energy represents the driving force for NVT systems, the thermodynamic potential that measures the “useful” work obtainable from a closed thermodynamic system at a constant temperature and volume.

GG density is given as:

\(exp(-y^\alpha)*\alpha*y^(\alpha*\delta - 1)/(scale*\gamma(\delta))\)

(see (Wikipedia))

A list of the member of the GG distribution family with the corresponding parameter settings is provided in Table 1 from reference (1). For example, to compute the HFE of a Weibull distribution model, we just set: \(alpha > 0\) and \(delta = 1\) (notice that parameter are named different in reference (1)).

helmholtz_free_energy(model, R = 8.31446261815324, T.abs = 310.15, ...)

# S4 method for missingORNULL
helmholtz_free_energy(
  model,
  R = 8.31446261815324,
  T.abs = 310.15,
  pars,
  log.base = exp(1),
  terms = FALSE
)

# S4 method for cdfMODEL
helmholtz_free_energy(
  model,
  R = 8.31446261815324,
  T.abs = 310.15,
  log.base = exp(1),
  terms = FALSE
)

# S4 method for cdfMODELlist
helmholtz_free_energy(
  model,
  R = 8.31446261815324,
  T.abs = 310.15,
  log.base = exp(1),
  terms = FALSE
)

# S4 method for ProbDistrList
helmholtz_free_energy(
  model,
  R = 8.31446261815324,
  T.abs = 310.15,
  log.base = exp(1),
  terms = FALSE
)

Arguments

model

An object from any of the classes created in MethylIT pipeline: cdfMODEL, cdfMODELlist, or ProbDistrList. If given, then the parameter values are taken from the model.

R

A number or NULL. The gas constant is given as default values, which is proportionality constant that relates the energy scale in physics to the temperature scale and the scale used for amount of substance.

T.abs

Absolute temperature (in Kelvin units). Default is T.abs = 310.15.

pars

Optional. A numerical vector containing the model parameter values in the given in order: alpha, scale, and delta.

log.base

A positive number. The same as in function log.

terms

logical(1). If term = TRUE, then a numerical vector with terms contributing to the Helmholtz free energy are provided.

Value

Helmholtz free energy of system according to the model. If R is the gas constant and T.abs given in Kelvin units, then the energy is returned in Joule per mol units.

Details

The values of the constants R and T.abs can be simply 1, which implies dimensionless units units, with perhaps no physical interpretation, only carrying informational meaning, i.e., only meaningful in the framework of information theory.

References

Crooks, Gavin E. (2015) The Amoroso Distribution. arXiv:1005.3274v2.

See also

Examples

## Loading the probability distribution models
data(gof, "MethylIT")
#> Warning: data set ‘MethylIT’ not found
## Gibb entropy at 310.15 K (Kelvin) in J * mol^-1
helmholtz_free_energy (gof)
#>        C1        C2        C3        T1        T2        T3 
#> -10162.67 -10103.27 -10131.46 -11493.86 -11523.32 -11641.73