Classification LDA model for simulated dataset of DMPs used in examples

This data/object carries the information about the classification performance of a Linear Discriminant ("LDA")) model on the set of dmps.

lda_perf

Format

lda_perf is list object consisting of the following elements:

Performance

Classification performance of the "LDA" model on the set of dmps

.
FDR

False discovery rate estimated for the "LDA" model on the set of dmps

model

The lda model fitted on the set of dmps

.

lda_perf is a list object obtained applying evaluateDIMPclass function on the set of dmps.