Weibull distribution with three parameters

Density, distribution function, quantile function and random generation for the Weibull distribution with three parameters

pweibull3P(q, shape = 1, scale = 1, mu = 0)

Arguments

q

vector of quantiles

shape

shape parameter, or slope, defaulting to 1

scale

scale parameter, or characteristic life, defaulting to 1

mu

location parameter, or failure free life, defaulting to 0

Value

3 parameters Weibull distribution

Examples

num.samples <- 10000
shape <- 0.75
scale <- 1
x <- rweibull(num.samples, shape = shape, scale = scale)
wei.model <- weibull3P(x)
#> 
#> *** Non-linear regression 
#> *** Trying nonlinear fit a Weibull 2P model...
#> *** Trying nonlinear fit a Weibull 3P model ...
#> *** Weibull-3P is the best fitted model
#> *** Performing nonlinear regression model crossvalidation...
y <- pweibull3P(x,
                shape = as.numeric(wei.model$Estimate[1]),
                scale = as.numeric(wei.model$Estimate[2]),
                mu = as.numeric(wei.model$Estimate[3]) )