Simple Cutpoint estimation

Internal function to estimate simple cutpoint according to Youden Index.

simpleCutPoint(
  LR,
  control.names,
  treatment.names,
  column,
  div.col,
  tv.col = NULL,
  tv.cut,
  clas.perf,
  classifier,
  interactions = NULL,
  n.pc,
  prop,
  post.cut = 0.5,
  maxnodes = NULL,
  ntree = 400,
  nsplit = 1L,
  cutp_data = FALSE,
  num.cores,
  tasks,
  ...
)

Arguments

LR, res, control.names, treatment.names, column, div.col

Same as in estimateCutPoint

column, div.col, tv.col, tv.cut, clas.perf

Same as in estimateCutPoint

classifier, n.pc, prop, cutp_data, num.cores, tasks

Same as in estimateCutPoint

interactions

If a logistic classifier is used. Variable interactions to consider in a logistic regression model. Any pairwise combination of the variable 'hdiv', 'TV', 'wprob', and 'pos' can be provided. For example: 'hdiv:TV', 'wprob:pos', 'wprob:TV', etc.

Value

Specified in function estimateCutPoint for parameter setting simple = TRUE

Details

This function is called by function estimateCutPoint.

Examples

## Get a set of potential DMPS (PS)
data(PS, package = 'MethylIT')

cutp <- simpleCutPoint(LR = PS, column = c(hdiv = TRUE, TV = TRUE,
                                           wprob = TRUE, pos = TRUE),
                       classifier = 'qda', n.pc = 4,
                       control.names = c('C1', 'C2', 'C3'),
                       treatment.names = c('T1', 'T2', 'T3'),
                       tv.cut = 0.5, clas.perf = TRUE, prop = 0.6,
                       div.col = 9L)

cutp
#> Cutpoint estimation with 'Youden Index' 
#> Simple cutpoint estimation 
#> Cutpoint = 117.04 
#> 
#> Cytosine sites from treatment have divergence values >= 117.04 
#> 
#> The accessible objects in the output list are: 
#>                     Length Class           Mode     
#> cutpoint             1     -none-          numeric  
#> testSetPerformance   6     confusionMatrix list     
#> testSetModel.FDR     1     -none-          numeric  
#> model               13     qdaDMP          list     
#> modelConfMatrix      6     confusionMatrix list     
#> initModel            1     -none-          character
#> postProbCut          1     -none-          logical  
#> postCut              1     -none-          logical  
#> classifier           1     -none-          character
#> statistic            1     -none-          logical  
#> optStatVal           1     -none-          logical  
#> cutpData             1     -none-          logical  
#> initModelConfMatrix  6     confusionMatrix list