Generates intervals for a GRanges objects

slidingGRegions generates sliding intervals of a specified width.

slidingGRegions(
  GR,
  win.size = 1,
  step.size = 1,
  num.cores = 1L,
  tasks = 0,
  verbose = FALSE,
  ...
)

Arguments

GR

A GRanges-class object or a list of GRanges-class objects or an object that can be coerced to a list of GRanges-class objects.

win.size

An integer. The size of the windows/intervals genomics.

step.size

Interval at which the regions/windows must be defined

num.cores, tasks

Parameters for parallel computation using package BiocParallel-package: the number of cores to use, i.e. at most how many child processes will be run simultaneously (see bplapply and the number of tasks per job (only for Linux OS).

verbose

Logical. Default is TRUE. If TRUE, then the progress of the computational tasks is given.

...

Not in use.

Value

An object from IRangesList-class.

Details

This function split a GRange object into intervals regions of fixed size. If win.size == step.size, then non-overlapping windows are obtained. slidingGRegions function generates sliding windows within each range of IRL, according to width and step, returning a IRangesList-class. If the sliding windows do not exactly cover a range in IRL, the last window is partial.

Author

Robersy Sanchez (https://genomaths.com).