Unique GRanges-class of methylation read counts filtered by coverage.

Given two GRanges-class objects, samples '1' and '2', carrying the counts of methylated (mC) and unmethylated (uC) cytosines in their metacolumns, this function will filter by coverage each cytosine site from each GRanges-class object.

uniqueGRfilterByCov(
  x,
  y = NULL,
  min.coverage = 4,
  min.meth = 0,
  min.umeth = 0,
  min.sitecov = 4,
  percentile = 0.9999,
  high.coverage = NULL,
  columns = c(mC = 1, uC = 2),
  ignore.strand = FALSE,
  num.cores = 1L,
  tasks = 0L,
  verbose = TRUE,
  ...
)

Arguments

x

An object from the classes 'GRanges', 'InfDiv', or 'pDMP' with methylated and unmethylated counts in its meta-column. If the argument 'y' is not given, then it is assumed that the first four columns of the GRanges-class metadata 'x' are counts: methylated and unmethylated counts for samples '1' and '2'. If x is a 'InfDiv', or 'pDMP', then 'y' is not needed, since samples '1' and '2' are the first four columns of these objects.

y

A GRanges-class or a GRangesList-class object with methylated and unmethylated counts in its meta-column. Default is NULL. If 'y' is a GRangesList-class object, then the first two metacolumns from 'x' are used in the pair-wise filtering toguether with the first two metacolumns from each GRanges-class element from 'y'.

min.coverage

An integer or an integer vector of length 2. Cytosine sites where the coverage in both samples, 'x' and 'y', are less than 'min.coverage' are discarded. The cytosine site is preserved, however, if the coverage is greater than 'min.coverage' in at least one sample. If 'min.coverage' is an integer vector, then the corresponding min coverage is applied to each sample.

min.meth

An integer or an integer vector of length 2. Cytosine sites where the numbers of read counts of methylated cytosine in both samples, '1' and '2', are less than 'min.meth' are discarded. If 'min.meth' is an integer vector, then the corresponding min number of reads is applied to each sample.

min.umeth

An integer or an integer vector of length 2. Min number of reads to consider cytosine position. Specifically cytosine positions where \(uC \leq min.umeth and mC > 0 and mC < min.meth\) hold will be removed, where mC and uC stand for the numbers of methylated and unmethylated reads. Default is \(min.umeth = 0\).

min.sitecov

An integer. The minimum total coverage. Only sites where the total coverage \(cov1 + cov2\) is greater than 'min.sitecov' are considered for downstream analysis, where cov1 and cov2 are the coverages for samples 1 and 2, respectively.

percentile

Threshold to remove the outliers (PCR bias) from each file and all files stacked. If 'high.coverage = NULL', then the threshold \(q\) will be computed as: $$q1 = quantile(cov1, probs=percentile)$$ $$q2 = quantile(cov2, probs=percentile)$$ $$q = min(q1, q2)$$

where \eqn{cov1} and \eqn{cov2} are the coverage vectors from samples 1
and 2, respectively.

high.coverage

An integer for read counts. Cytosine sites having higher coverage than this are discarded. Default is NULL. If high.coverage is not NULL, then the percentile argument is disregarded and high.coverage is used as threshold to remove the PCR bias. If percentile is not null, then q = max(q, high.coverage).

columns

Vector of integer numbers of the columns (from each GRanges meta-column) where the methylated and unmethylated counts are provided. If not provided, then the methylated and unmethylated counts are assumed to be at columns 1 and 2, respectively.

ignore.strand

When set to TRUE, the strand information is ignored in the overlap of GRanges-class objects. This is a parameter passed to uniqueGRanges function. Default value: FALSE.

num.cores, tasks

Parameters for parallel computation using package BiocParallel-package: the number of cores to use, i.e. at most how many child processes will be run simultaneously (see bplapply and the number of tasks per job (only for Linux OS).

verbose

if TRUE, prints the function log to stdout

...

Additional parameters for uniqueGRanges function.

Value

if 'x' and 'y' are GRanges-class object, then a GRanges-class with the columns of methylated and unmethylated counts filtered for each cytosine position. A GRangesList-class object will returned, if 'y' is a GRangesList-class object of same length as length(y) and named as names(y).

Details

Cytosine sites with 'coverage' > 'min.coverage' in at least one of the samples are preserved. Positions with 'coverage' < 'min.coverage' in both samples, 'x' and 'y', are removed. Positions with 'coverage' > 'percentile' (e.g., 99.9 percentile) are removed as well. It is expected that the columns of methylated and unmethylated counts are given.

This function is addressed to create pair-wise GRanges-class object with four metacolumns of count: samples '1' and '2', carrying the counts of methylated (mC) and unmethylated (uC) cytosines in their metacolumns, respectively. Counts from sample 1 are typically used as reference counts in computing information divergences in the downtstream analysis.

Examples

### Create new data
df1 <- data.frame(chr = 'chr1', start = 11:16, end = 11:16,
                mC = c(2,10,7,9,1,10), uC = c(30,20,4,8,0,10))

df2 <- data.frame(chr = 'chr1', start = 12:18, end = 12:18,
                mC2 = 1:7, uC2 = 0:6)

gr1 <- makeGRangesFromDataFrame(df1, keep.extra.columns = TRUE)
gr2 <- makeGRangesFromDataFrame(df2, keep.extra.columns = TRUE)

## Filtering
r1 <- uniqueGRfilterByCov(gr1, gr2, ignore.strand = TRUE)
#>  *** Building coordinates for the new GRanges object ...
#>  *** Setting strand information to * ...
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#>  *** Sorting by chromosomes and start positions...
r1
#> GRanges object with 3 ranges and 4 metadata columns:
#>       seqnames    ranges strand |        mC        uC       mC2       uC2
#>          <Rle> <IRanges>  <Rle> | <numeric> <numeric> <numeric> <numeric>
#>   [1]     chr1        13      * |         7         4         2         1
#>   [2]     chr1        15      * |         1         0         4         3
#>   [3]     chr1        17      * |         0         0         6         5
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## Cytosine position with coordinate 15 (rows #2) can pass the
## filtering conditions of min.coverage = 4 and lead to meaningless
## situations with methylation levels p = 1/(1 + 0) = 1
r1[2]
#> GRanges object with 1 range and 4 metadata columns:
#>       seqnames    ranges strand |        mC        uC       mC2       uC2
#>          <Rle> <IRanges>  <Rle> | <numeric> <numeric> <numeric> <numeric>
#>   [1]     chr1        15      * |         1         0         4         3
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## The last situation can be prevent, in this case, by setting
## min.meth = 1:
r1 <- uniqueGRfilterByCov(gr1, gr2, min.meth = 1, ignore.strand = TRUE)
#>  *** Building coordinates for the new GRanges object ...
#>  *** Setting strand information to * ...
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#>  *** Sorting by chromosomes and start positions...
r1
#> GRanges object with 2 ranges and 4 metadata columns:
#>       seqnames    ranges strand |        mC        uC       mC2       uC2
#>          <Rle> <IRanges>  <Rle> | <numeric> <numeric> <numeric> <numeric>
#>   [1]     chr1        13      * |         7         4         2         1
#>   [2]     chr1        17      * |         0         0         6         5
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths