DMP clustering

Given a 'pDMP' object carrying DMPs obtained in Methyl-IT downstream analysis, function 'dmpClusters' build clusters of DMPs, which can be further tested to identify differentially methylated regions (DMRs) with countTest2 function.

dmpClusters(
  GR,
  maxDist = 3,
  minNumDMPs = 1,
  maxClustDist = NULL,
  method = c("relaxed", "fixed.int"),
  chromosomes = NULL,
  ignore.strand = TRUE,
  num.cores = 1L,
  tasks = 0L,
  verbose = TRUE,
  ...
)

Arguments

GR
An object from 'pDMP' class, which is returned by selectDIMP function.

maxDist
maximum distance at which two reported bases sites from the same cluster can be separated. Default: \(maxDist = 3\).

minNumDMPs
Minimum number of DMPs inside of each cluster. Default: \(minNumDMPs = 1\).

maxClustDist
Clusters separated by a distance lesser than 'maxClustDist' positions are merged. Default: \(maxClustDist = NULL\). If \(0 < maxClustDist < maxDist\), then maxClustDist will be recalculated as \(maxClustDist = maxDist + 1\).

method
Two different approaches are implemented to clustering DMPs:

"relaxed":

DMP ranges which are separated by a distance less than 'maxClustDist' are merged and ranges with less than 'minNumDMPs' are removed.

"fixed.int":

It will gnetate a partition where the distance between consecutive DMPs is not greater than 'maxDist'. Ranges with less than 'minNumDMPs' are removed. If, aditionally, a value maxClustDist > 0 is provided, then the "relaxed" approach is applied to the ranges from the first step.

chromosomes
vector of characters labeling the chromosomes included in the analysis. Default: chromosomes = NULL (all chromosomes are included).

ignore.strand
Same as in findOverlaps-methods.

num.cores, tasks
integer(1). The number of cores to use, i.e. at most how many child processes will be run simultaneously (see bplapply function from BiocParallel package).The number of tasks per job. value must be a scalar integer >= 0L (see MulticoreParam from BiocParallel package).

verbose
if TRUE, prints the function log to stdout.

...
Further parameters for uniqueGRanges function.

Value

A GRanges object with the numbers of positions inside each cluster, where DMPs were reported in at least one of the samples.

Details

Two algorithmic approaches are implemented, named: "relaxed" and "fixed.int" (see the decription of parmeter 'method'). The "fixed.int" is mostly addressed to find specific methylation patterns, but the price is the number of DMRs found is lower.

The number of DMPs reported in each cluster corresponds to the numbers of sites inside the cluster where DMPs were found in at least one of the samples (from control or from treatment). That is, dmpClusters is only a tool to locate regions with high density of DMPs from all the samples. It does not detect DMRs. It is assumed that only DMP coordinates are given in the 'GR' object. That is, all the sites provided are considered in the computation.

References

Sanchez R, Mackenzie SA (2016) Information Thermodynamics of Cytosine DNA Methylation. PLOS ONE 11(3): e0150427. https://doi.org/10.1371/journal.pone.0150427

Examples

## Get a dataset of dmps from the package
data(dmps)

## Build clusters of DMPs taking into account the DNA strand
x1 = dmpClusters2(GR = dmps, maxDist = 7, minNumDMPs = 6,
                 method = "fixed.int", ignore.strand = FALSE,
                 verbose = FALSE)

#> Error in dmpClusters2(GR = dmps, maxDist = 7, minNumDMPs = 6, method = "fixed.int",     ignore.strand = FALSE, verbose = FALSE): could not find function "dmpClusters2"

data.frame(x1)

#> Error in data.frame(x1): object 'x1' not found

# \donttest{
## Build clusters of DMPs ignoring DNA strand and maxClustDist = 7
x2 = dmpClusters2(GR = dmps, maxDist = 7, minNumDMPs = 6,
                  maxClustDist = 7, method = "fixed.int",
                  num.cores=2L, ignore.strand = TRUE,
                  verbose = FALSE)

#> Error in dmpClusters2(GR = dmps, maxDist = 7, minNumDMPs = 6, maxClustDist = 7,     method = "fixed.int", num.cores = 2L, ignore.strand = TRUE,     verbose = FALSE): could not find function "dmpClusters2"

DataFrame(data.frame(x2))

#> Error in data.frame(x2): object 'x2' not found

## The relaxed approach with method = "relaxed"
x3 = dmpClusters2(GR = dmps, minNumDMPs = 6, method = "relaxed",
                  maxClustDist = 10, ignore.strand = TRUE,
                  verbose = FALSE)

#> Error in dmpClusters2(GR = dmps, minNumDMPs = 6, method = "relaxed", maxClustDist = 10,     ignore.strand = TRUE, verbose = FALSE): could not find function "dmpClusters2"

DataFrame(data.frame(x3))

#> Error in data.frame(x3): object 'x3' not found

## ==== Setting up the experiment design to test for DMRs ===
nams <- names(dmps)
dmps_at_clusters <- getDMPatRegions(GR = dmps, regions = x3,
                                    ignore.strand = TRUE)

#> Error in FUN(X[[i]], ...): object 'x3' not found

dmps_at_clusters <- uniqueGRanges(dmps_at_clusters, columns = 2L,
                                ignore.strand = TRUE, type = 'equal',
                                verbose = FALSE)

#> Error in is(ListOfGranges, "list"): object 'dmps_at_clusters' not found

colnames(mcols(dmps_at_clusters)) <- nams

#> Error in colnames(mcols(dmps_at_clusters)) <- nams: object 'dmps_at_clusters' not found

colData <- data.frame(condition = factor(c('CT', 'CT', 'CT',
                                        'TT', 'TT', 'TT'),
                                        levels = c('CT', 'TT')),
                    nams, row.names = 2)

## Build a RangedGlmDataSet object
ds <- glmDataSet(GR = dmps_at_clusters, colData = colData)

#> Error in glmDataSet(GR = dmps_at_clusters, colData = colData): object 'dmps_at_clusters' not found

## ================ Testing to detect DMRs ===========
dmrs <- countTest2(DS = ds, num.cores = 4L, minCountPerIndv = 4,
                maxGrpCV = c(1, 1), Minlog2FC = 0.5,
                CountPerBp = 0.001, test = 'LRT',
                verbose = TRUE)

#> *** Number of GR after filtering counts 92

#> *** GLM...

dmrs

#> GRanges object with 22 ranges and 12 metadata columns:
#>           seqnames        ranges strand |        A1        A2        B1
#>              <Rle>     <IRanges>  <Rle> | <integer> <integer> <integer>
#>    gene22     chr1   42001-43001      * |        44        53       187
#>    gene26     chr1   50001-51001      * |       174       139        32
#>    gene29     chr1   56001-57001      * |        12        19       129
#>    gene44     chr1   86001-87001      * |       110       173        23
#>    gene57     chr1 112001-113001      * |        25        94       187
#>       ...      ...           ...    ... .       ...       ...       ...
#>   gene106     chr2 210001-211001      * |       122        85         3
#>   gene107     chr2 212001-213001      * |       181       111        76
#>   gene113     chr2 224001-225001      * |       142       172        59
#>   gene114     chr2 226001-227001      * |        33        34       111
#>   gene117     chr2 232001-233001      * |        55        77        13
#>                  B2             log2FC  scaled.deviance               pvalue
#>           <integer>          <numeric>        <numeric>            <numeric>
#>    gene22        85   1.03109108779261 11.5705078723797 0.000670061144147836
#>    gene26        62  -1.20290840827015 30.9662853450191 2.62549774873693e-08
#>    gene29        64   1.82870298441974 29.4987592669987 5.59525099643946e-08
#>    gene44        88 -0.935916696330904  3.9493596280495   0.0468892079882497
#>    gene57       174   1.10975446522135 6.43211707492669    0.011207491350226
#>       ...       ...                ...              ...                  ...
#>   gene106        16  -2.38827981409893 30.9821950561452 2.60406378697818e-08
#>   gene107        21   -1.1020428237649 5.75602794152791   0.0164321785717638
#>   gene113       127 -0.523646312195051 3.84537303089257   0.0498834241817064
#>   gene114       158   1.39001876021087 15.1218903312962 0.000100787798473054
#>   gene117        43 -0.857450231851222 4.67402216233158   0.0306221309485589
#>                    model             adj.pval   CT.SignalDensity
#>              <character>            <numeric>          <numeric>
#>    gene22   Neg.Binomial   0.0012284454309377 0.0484515484515484
#>    gene26   Neg.Binomial 1.92536501574041e-07  0.156343656343656
#>    gene29   Neg.Binomial 3.07738804804171e-07 0.0154845154845155
#>    gene44   Neg.Binomial   0.0491220274162616  0.141358641358641
#>    gene57   Neg.Binomial   0.0164376539803314 0.0594405594405594
#>       ...            ...                  ...                ...
#>   gene106   Neg.Binomial 1.92536501574041e-07  0.103396603396603
#>   gene107   Neg.Binomial   0.0212651722693414  0.145854145854146
#>   gene113   Neg.Binomial   0.0498834241817064  0.156843156843157
#>   gene114 Neg.Binomial.W 0.000256868046468425 0.0334665334665335
#>   gene117   Neg.Binomial   0.0374270489371276 0.0659340659340659
#>              TT.SignalDensity SignalDensityVariation
#>                     <numeric>              <numeric>
#>    gene22   0.135864135864136     0.0874125874125874
#>    gene26   0.046953046953047     -0.109390609390609
#>    gene29  0.0964035964035964     0.0809190809190809
#>    gene44  0.0554445554445554    -0.0859140859140859
#>    gene57    0.18031968031968      0.120879120879121
#>       ...                 ...                    ...
#>   gene106 0.00949050949050949    -0.0939060939060939
#>   gene107  0.0484515484515484    -0.0974025974025974
#>   gene113  0.0929070929070929    -0.0639360639360639
#>   gene114   0.134365634365634      0.100899100899101
#>   gene117   0.027972027972028     -0.037962037962038
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths

# }