DMP cluster builder

Given a 'pDMP' object carrying DMPs obtained in Methyl-IT downstream analysis, function "dmpClusters" build clusters of DMPs, which can be further tested to identify differentially methylated regions (DMRs) with countTest2 function.

dmpClusters(
  GR,
  maxDist = 3,
  minNumDMPs = 1,
  chromosomes = NULL,
  num.cores = 1L,
  tasks = 0L,
  verbose = TRUE,
  ...
)

Arguments

GR
An object from 'pDMP' class, which is returned by selectDIMP function.

maxDist
maximum distance at which two reported bases sites from the same cluster can be separated. Default: \(maxDist = 3\).

minNumDMPs
minimum number of marked bases inside of each cluster. Default: \(minNumDMPs = 1\).

chromosomes
vector of characters labeling the chromosomes included in the analysis. Default: chromosomes = NULL (all chromosomes are included).

num.cores, tasks
integer(1). The number of cores to use, i.e. at most how many child processes will be run simultaneously (see bplapply function from BiocParallel package).The number of tasks per job. value must be a scalar integer >= 0L (see MulticoreParam from BiocParallel package).

verbose
if TRUE, prints the function log to stdout.

...
Further parameters for uniqueGRanges function.

Value

GRanges object with the numbers of positions inside each cluster where DMPs were reported in at least one of the samples.

Details

DNA base coordinates provided in the GR object are used to build a binary string of zeros and ones . The binary string of zeros and ones is used in the detection of cluster of DMPs. Postulating that a reported DNA base is found at the beginning and the end of a cluster, genome-wide screening can be performed where two consecutive targeting base positions are separated by less than a given threshold maxDist bases (1). DMPs from all the samples included in GR are considered to build the clusters.

The number of DMPs reported in each cluster corresponds to the numbers of positions inside the cluster where DMPs were reported in at least one of the samples.

References

Sanchez R, Mackenzie SA (2016) Information Thermodynamics of Cytosine DNA Methylation. PLOS ONE 11(3): e0150427. https://doi.org/10.1371/journal.pone.0150427

Examples

## Get a dataset of potential signals and the estimated cutpoint from the
## package and performs cutpoint estimation
data(PS, cutpoint)

## The estimated cutpoints are used to discriminate signals from the noise.
## That is, DMPs are selected using the cupoints
dmps <- selectDIMP(PS, div.col = 9L, cutpoint = cutpoint$cutpoint)

## Build clusters of DMPs
x1 = dmpClusters(GR = dmps, maxDist = 7, minNumDMPs = 6,num.cores=2L,
                verbose = FALSE)
x1

#> GRanges object with 60 ranges and 1 metadata column:
#>        seqnames    ranges strand |      dmps
#>           <Rle> <IRanges>  <Rle> | <numeric>
#>    [1]        1    91-119      * |         7
#>    [2]        1   208-234      * |         8
#>    [3]        1   552-575      * |         7
#>    [4]        1   765-786      * |         9
#>    [5]        1 1125-1142      * |         6
#>    ...      ...       ...    ... .       ...
#>   [56]        1 9409-9427      * |         6
#>   [57]        1 9444-9469      * |         8
#>   [58]        1 9538-9559      * |         8
#>   [59]        1 9726-9742      * |         6
#>   [60]        1 9911-9958      * |        14
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## ==== Setting up the experimennt design to test for DMRs ===
nams <- names(dmps)
dmps_at_clusters <- getDIMPatGenes(GR = dmps, GENES = x1,
                                    ignore.strand = TRUE)
dmps_at_clusters <- uniqueGRanges(dmps_at_clusters, columns = 2L,
                                    ignore.strand = TRUE, type = "equal",
                                    verbose = FALSE)
colnames(mcols(dmps_at_clusters)) <- nams

colData <- data.frame(condition = factor(c("CT", "CT", "CT",
                                           "TT", "TT", "TT"),
                                         levels = c("CT", "TT")),
                      nams,
                      row.names = 2)

## Build RangedGlmDataSet is created
ds <- glmDataSet(GR = dmps_at_clusters, colData = colData)

## ================ Testing for DMRs ===========
dmrs <- countTest2(DS = ds, num.cores = 4L,
                   minCountPerIndv = 4,
                   maxGrpCV = c(1, 1),
                   Minlog2FC = 0.5,
                   CountPerBp = 0.001,
                   test = "LRT",
                   verbose = TRUE)

#> *** Number of GR after filtering counts 6

#> *** GLM...

dmrs

#> GRanges object with 6 ranges and 14 metadata columns:
#>       seqnames    ranges strand |        C1        C2        C3        T1
#>          <Rle> <IRanges>  <Rle> | <numeric> <numeric> <numeric> <numeric>
#>   [1]        1 1204-1238      * |         0         0         0         5
#>   [2]        1 2142-2172      * |         0         0         0         2
#>   [3]        1 2325-2359      * |         0         0         1         5
#>   [4]        1 4656-4708      * |         0         2         0         8
#>   [5]        1 7722-7775      * |         1         0         0         5
#>   [6]        1 9911-9958      * |         0         0         0         4
#>              T2        T3           log2FC  scaled.deviance              pvalue
#>       <numeric> <numeric>        <numeric>        <numeric>           <numeric>
#>   [1]         5         4 1.73460105538811 8.48613318671429 0.00357863605668028
#>   [2]         6         4  1.6094379124341 6.94832058758156 0.00838979577447831
#>   [3]         3         5 1.38629436111939  30.839145590753 0.00514571648076653
#>   [4]         5         6 1.48160454025156 20.6669445813151  0.0104485974536701
#>   [5]         1         7 1.38629436139456  7.7096310654329  0.0499909384290531
#>   [6]         9         3 1.84582669049833 9.55203643495998 0.00199727710868162
#>                model           adj.pval    CT.SignalDensity   TT.SignalDensity
#>             <factor>          <numeric>           <numeric>          <numeric>
#>   [1] Neg.Binomial.W 0.0102914329615331                   0  0.133333333333333
#>   [2] Neg.Binomial.W 0.0125383169444042                   0  0.129032258064516
#>   [3]   QuasiPoisson 0.0102914329615331 0.00952380952380952  0.123809523809524
#>   [4]   QuasiPoisson 0.0125383169444042  0.0125786163522013  0.119496855345912
#>   [5]   QuasiPoisson 0.0499909384290531 0.00617283950617284 0.0802469135802469
#>   [6] Neg.Binomial.W 0.0102914329615331                   0  0.111111111111111
#>       SignalDensityVariation
#>                    <numeric>
#>   [1]      0.133333333333333
#>   [2]      0.129032258064516
#>   [3]      0.114285714285714
#>   [4]      0.106918238993711
#>   [5]     0.0740740740740741
#>   [6]      0.111111111111111
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths