Information divergence estimator

The Information divergence of methylation levels is estimated using the direct estimation or a Bayesian approach of the methylation levels. Hellinger divergence is computed as given in reference 1.

estimateBayesianDivergence(
  x,
  Bayesian = FALSE,
  JD = FALSE,
  num.cores = 1,
  tasks = 0L,
  columns = c(mC1 = 1, uC1 = 2, mC2 = 3, uC2 = 4),
  meth.level = FALSE,
  preserve.gr = FALSE,
  logbase = 2,
  verbose = TRUE
)

Arguments

x
A matrix of counts or GRanges object with the table of counts in the meta-columns (methylated mC and unmethylated uC cytosines). Unless specified in the parameter 'columns', the methylation counts must be given in the first four columns: "mC1" and "uC1" methylated and unmethylated counts for control sample, and "mC2" and "uC2" methylated and unmethylated counts for treatment sample, respectively.

Bayesian
logical. Whether to perform the estimations based on posterior estimations of methylation levels.

JD
Logic (Default:FALSE). Option on whether to add a column with values of J-information divergence (see estimateJDiv). It is only compute if JD = TRUE and meth.level = FALSE.

num.cores
The number of cores to use, i.e. at most how many child processes will be run simultaneously (see 'bplapply' function from BiocParallel package).

tasks
integer(1). The number of tasks per job. value must be a scalar integer >= 0L. In this documentation a job is defined as a single call to a function, such as bplapply, bpmapply etc. A task is the division of the X argument into chunks. When tasks == 0 (default), X is divided as evenly as possible over the number of workers (see MulticoreParam from BiocParallel package).

columns
Vector of integer numbers of the columns where the counts "mC1", "uC1", "mC2", and "uC2" are in the matrix (default 1 to 4). That is, the input could have more than 4 columns, but only 4 columns with the counts are used.

meth.level
methylation levels can be provided in place of counts.

preserve.gr
Logic (Default:FALSE). Option of whether to preserve all the metadata from the original GRange object.

logbase
Logarithm base used to compute the JD (if JD = TRUE). Logarithm base 2 is used as default (bit unit). Use logbase = exp(1) for natural logarithm.

verbose
if TRUE, prints the function log to stdout

Value

The input matrix or GRanges object with the four columns of counts and additional columns. If Bayessian = TRUE, the results are based on the posterior estimations of methylation levels. 1) p1" and "p2": methylation levels for control and treatment; 2) "hdiv": Hellinger divergence; 3) "bay.TV". "TV": total variation TV = p2 - p1 is based on simple counts

Details

For the current version, the Information divergence of methylation levels is estimated based on Hellinger divergence. If read counts are provided, then Hellinger divergence is computed as given in the first formula from Theorem 1 from reference 1. In the present case,

hdiv = 2*(n[1] + 1)*(n[2] + 1)*((sqrt(p[1]) - sqrt(p[2]))^2 + (sqrt(1 -
p[1]) - sqrt(1 - p[2]))^2)/(n[1] + n[2] + 2)

where n[1] and n[2] are the coverage for the control and treatment,
respectively. Notice that each row from the matrix of counts correspond
to a single cytosine position and has four values corresponding to "mC1"
and "uC1" (control), and mC2" and "uC2" for treatment.

If the methylation levels are provided in place of counts, then
Hellinger divergence is computed as:
hdiv = (sqrt(p[1]) - sqrt(p[2]))^2 + (sqrt(1 -p[1]) - sqrt(1 - p[2]))^2

This formula assumes that the probability vectors derived from the
methylation levels (p_ij) p_j = c(p_ij, 1 - p_ij) (see function
'estimateHellingerDiv') are an unbiased estimation of the expected one.

References

Basu A., Mandal A., Pardo L (2010) Hypothesis testing for two discrete populations based on the Hellinger distance. Stat Probab Lett 80: 206-214.

Examples

## The read count data are created
    x <- data.frame(chr = "chr1", start = 1:10,
                    end = 1:10,strand = '*',
                    mC1 = rnbinom(size = 10, mu = 4, n = 500),
                    uC1 = rnbinom(size = 10, mu = 4, n = 500),
                    mC2 = rnbinom(size = 10, mu = 4, n = 500),
                    uC2 = rnbinom(size = 10, mu = 4, n = 500))
    x <- makeGRangesFromDataFrame(x, keep.extra.columns = TRUE)
## Estimation of the information divergences
    hd <- estimateBayesianDivergence(x, JD = TRUE)

#> *** Estimating Hellinger divergence... 
#> * Coercing from list to vector...

## Keep in mind that Hellinger and J divergences are, in general, correlated!
    cor.test(x = as.numeric(hd$hdiv), y = as.numeric(hd$jdiv),
            method = "kendall")

#> 
#> 	Kendall's rank correlation tau
#> 
#> data:  as.numeric(hd$hdiv) and as.numeric(hd$jdiv)
#> z = 27.624, p-value < 2.2e-16
#> alternative hypothesis: true tau is not equal to 0
#> sample estimates:
#>       tau 
#> 0.8306315 
#>