J Information Divergence of Methylation Levels

Given a the methylation levels from two individuals at a given cytosine site, this function computes the J information divergence (JD) between methylation levels. The motivation to introduce JD in Methyl-IT is founded on:

It is a symmetrised form of Kullback–Leibler divergence (\(D_KL\)). Kullback and Leibler themselves actually defined the divergence as: \(D_KL(P || Q) + D_KL(Q || P)\), which is symmetric and nonnegative, where the probability distributions P and Q are defined on the same probability space (see Wikipedia).

In general, JD is highly correlated with Hellinger divergence, which is the main divergence currently used in Methyl-IT (see examples for function estimateDivergence.

By construction, the unit of measurement of JD is: bit of information, which set the basis for further information-thermodynamics analyses.

estimateJDiv(p, logbase = 2)

Arguments

p
A numerical vector of the methylation levels p = c(p1, p2) from individuals 1 and 2.

logbase
Logarithm base used to compute the JD. Logarithm base 2 is used as default. Use logbase = exp(1) for natural logarithm.

Value

The J divergence value for the given methylation levels is returned

Details

The methylation level \(p_ij\) at a given cytosine site i from an individual j corresponds to a probability vector \(q_ij = c(p_ij, 1 - p_ij)\). Then, the J-information divergence between the methylation levels p_i1 and p_i2 is the divergence between the vectors \(q_i1 = (p_i1, 1 - p_i1)\) and \(q_i2 = c(p_i2, 1 - p_i2)\). JD is computed as in reference (1):

$$JD(q_i1, q_i2) = (q_i1[1] * log(q_i1[1]/q_i2[1]) + q_i1[2] * log(q_i1[2]/q_i2[2]) + q_i2[1] * log(q_i2[1]/q_i1[1]) + q_i2[2] * log(q_i2[2]/q_i1[2]))/2$$.

References

Lin J. Divergence Measures Based on the Shannon Entropy. IEEE Trans Inform Theory, 1991, 37:145–51.

Sanchez R, Mackenzie SA. Information thermodynamics of cytosine DNA methylation. PLoS One, 2016, 11:e0150427.

See also

Examples

p <- c(0.5, 0.5)
estimateJDiv(p)

#> [1] 0

## A numerical trick is implemented. The J-divergence values are the
## same for the following vectors:
p <- c(0.9999999999999999, 0)
q <- c(1, 0)

estimateJDiv(p) == estimateJDiv(q)

#> [1] TRUE