J Information Divergence of Methylation Levels

Given a the methylation levels from two individuals at a given cytosine site, this function computes the J information divergence (JD) between methylation levels. The motivation to introduce JD in Methyl-IT is founded on:

It is a symmetrised form of Kullback–Leibler divergence (\(D_{KL}\)). Kullback and Leibler themselves actually defined the divergence as: \(D_{KL}(P || Q) + D_{KL}(Q || P)\), which is symmetric and nonnegative, where the probability distributions P and Q are defined on the same probability space (see Wikipedia).

In general, JD is highly correlated with Hellinger divergence, which is the main divergence currently used in Methyl-IT (see examples for function estimateDivergence.

By construction, the unit of measurement of JD is: bit of information, which set the basis for further information-thermodynamics analyses.

estimateJDiv(p, logbase = 2)

Arguments

p
A numerical vector of the methylation levels p = c(p1, p2) from individuals 1 and 2.

logbase
Logarithm base used to compute the JD. Logarithm base 2 is used as default. Use logbase = exp(1) for natural logarithm.

Value

The J divergence value for the given methylation levels is returned

Details

The methylation levels \(p_{ij}\), at a given cytosine site i from an individual j, lead to the probability vectors \(p^{ij} = (p_{ij}, 1 - p_{ij})\). Then, the J-information divergence between the methylation levels \(p_{ij}\) and the methylation levels \(q_i\), used as reference, corresponds to the divergence between the probability vectors \(p^{ij}\) and \(q^i = (q_i, 1 - q_i)\). JD is computed as in reference (1-2):

$$JD(p^{ij}, q^i) = ((p_{ij} - q_i) * log(p_{ij}/q_i) + (q_i - p_{ij}) * log((1 - p_{ij})/(1 - q_i)))/2$$

References

J. K. Chung, P. L. Kannappan, C. T. Ng, P. K. Sahoo, Measures of distance between probability distributions. J. Math. Anal. Appl. 138, 280–292 (1989).

Lin J. Divergence Measures Based on the Shannon Entropy. IEEE Trans Inform Theory, 1991, 37:145–51.

Sanchez R, Mackenzie SA. Information thermodynamics of cytosine DNA methylation. PLoS One, 2016, 11:e0150427.

See also

Examples

p <- c(0.5, 0.5)
estimateJDiv(p)

#> [1] 0

## A numerical trick is implemented. The J-divergence values are the
## same for the following vectors:
p <- c(0.9999999999999999, 0)
q <- c(1, 0)

estimateJDiv(p) == estimateJDiv(q)

#> [1] FALSE