Generalized Gamma distribution

Probability density function (PDF), cumulative density function (CDF), quantile function and random generation for the Generalized Gamma (GG) distribution with 3 or 4 parameters: alpha, scale, mu, and psi. The function is reduced to GGamma distribution with 3 parameters by setting mu = 0.

dggamma(q, alpha = 1, scale = 1, mu = 0, psi = 1, log.p = FALSE)

pggamma(
  q,
  alpha = 1,
  scale = 1,
  mu = 0,
  psi = 1,
  lower.tail = TRUE,
  log.p = FALSE
)

qggamma(
  p,
  alpha = 1,
  scale = 1,
  mu = 0,
  psi = 1,
  lower.tail = TRUE,
  log.p = FALSE
)

rggamma(n, alpha = 1, scale = 1, mu = 0, psi = 1)

Arguments

q
numeric vector.

alpha
numerical parameter, strictly positive (default 1). The generalized gamma becomes the gamma distribution for alpha = 1.

scale, psi
the same real positive parameters as is used for the Gamma distribution. These are numerical and strictly positives; default 1. (see ?pgamma).

mu
location parameter (numerical, default 0).

log.p
logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail
logical; if TRUE (default), probabilities are \(P(X<=x)\), otherwise, \(P(X > x)\)

p
vector of probabilities.

n
number of observations.

Value

GG PDF values (3-parameters or 4-parameters) for dggamma, GG probability for pggamma, quantiles or GG random generated values for rggamma.

Details

Details about these function can be found in references 1 to 3. You may also see section Note at ?pgamma or ?rgamma. Herein, we are using Stacy' s formula (references 2 to 3) with the parametrization given in reference 4 (equation 6, page 12). As in the case of gamma distribution function, the cumulative distribution function (as given in equation 12, page 13 from reference 4) is expressed in terms of the lower incomplete gamma function (see ?pgamma).

The GG distribution with parameters \eqn{\alpha}, \eqn{\beta} (scale),
\eqn{\psi}, and \eqn{\mu} has density:

\deqn{f(x | \alpha, \beta, \mu, \psi) = \alpha exp(-((x-\mu)/
    \beta)^\alpha) ((x-\mu)/\beta)^(\alpha * \psi - 1)/(\beta Gamma(\psi))}

References

Handbook on STATISTICAL DISTRIBUTIONS for experimentalists (p. 73) by Christian Walck. Particle Physics Group Fysikum. University of Stockholm (e-mail: walck@physto.se )

Stacy, E. W. A Generalization of the Gamma Distribution. Ann. Math. Stat. 33, 1187–1192 (1962).

Stacy E, Mihram G (1965) Parameter estimation for a generalized gamma distribution. Technometrics 7: 349-358.

Sanchez, R. & Mackenzie, S. A. Information Thermodynamics of Cytosine DNA Methylation. PLoS One 11, e0150427 (2016).

Examples

q <- (1:9)/10
pggamma(q, alpha = 1, scale = 1, mu = 0,
        psi = 1, lower.tail = TRUE, log.p = FALSE)

#> [1] 0.09516258 0.18126925 0.25918178 0.32967995 0.39346934 0.45118836 0.50341470
#> [8] 0.55067104 0.59343034

## To fit random generated numbers
set.seed(123)
x <- rggamma(2000, alpha = 1.03, psi = 0.75, scale = 2.1)
fitGGammaDist(x)

#> *** Trying nonlinear fit to a generalized 3P Gamma distribution model  values...

#> *** Performing nonlinear regression model  crossvalidation...

#>        Estimate  Std. Error  t value Pr(>|t|))      Adj.R.Square
#> alpha 0.9673236 0.004522499 213.8914         0 0.999999987853079
#> scale 1.9144838 0.014958723 127.9844         0                  
#> psi   0.8336444 0.005740757 145.2151         0                  
#>                     rho       R.Cross.val                  DEV              AIC
#> alpha 0.999999987810436 0.999883383772831 2.02246091083126e-06 -16008.797746206
#> scale                                                                          
#> psi                                                                            
#>                     BIC     COV.alpha     COV.scale       COV.psi COV.mu    N
#> alpha -15986.3941363678  2.045299e-05  6.688155e-05 -2.580564e-05     NA 2000
#> scale                    6.688155e-05  2.237634e-04 -8.574600e-05     NA 2000
#> psi                     -2.580564e-05 -8.574600e-05  3.295629e-05     NA 2000
#>          model
#> alpha GGamma3P
#> scale         
#> psi