Quadratic Discriminant Analysis (QDA) using Principal Component Analysis (PCA)

The principal components (PCs) for predictor variables provided as input data are estimated and then the individual coordinates in the selected PCs are used as predictors in the qda

Predict using a PCA-LDA model built with function 'pcaLDA'

pcaQDA(
  formula = NULL,
  data = NULL,
  grouping = NULL,
  n.pc = 1,
  scale = FALSE,
  center = FALSE,
  tol = 1e-04,
  method = "moment",
  max.pc = NULL
)

# S3 method for pcaQDA
predict(
  object,
  newdata,
  type = c("qda.pred", "class", "posterior", "pca.ind.coord", "all"),
  ...
)

Arguments

formula
Same as in qda from package 'MASS'.

data
Same as in qda from package 'MASS'.

grouping
Same as in qda from package 'MASS'.

n.pc
Number of principal components to use in the qda.

scale
Same as in prcomp from package 'stats'.

center
Same as in prcomp from package 'stats'.

tol
Same as in prcomp from package 'stats'.

method
Same as in qda from package 'MASS'.

max.pc
Same as in parameter 'rank.' from prcomp from package 'stats'.

object
To use with function 'predict'. A 'pcaQDA' object containing a list of two objects: 1) an object of class inheriting from "qda" and 2) an object of class inheriting from "prcomp".

newdata
To use with function 'predict'. New data for classification prediction.

type
To use with function 'predict'. The type of prediction required. The default is "all" basic predictions: classes and posterior classification probabilities. Option "qda.pred" returns the object given by function 'predict.qda' from MASS package: 'class', 'posterior', 'scores' (cases scores on discriminant variables, see qda.

...
Not in use.

Value

Function 'pcaQDA' returns an object ("pcaQDA") consisting of a list with
two objects:
    1) 'qda': an object of class 'qda' from package 'MASS'.
    2) 'pca': an object of class 'prcomp' from package 'stats'.
For information on how to use these objects see ?qda and ?prcomp.

Details

The principal components (PCs) are obtained using the function 'prcomp' from R package 'stats', while the qda is performed using the 'qda' function from R package 'MASS'. The current application only uses basic functionalities of mentioned functions. As shown in the example, 'pcaQDA' function can be used in general classification problems.

See also

Examples

data(iris)
qd1 <- pcaQDA(formula = Species ~ Petal.Length + Sepal.Length + Sepal.Width,
            data = iris, n.pc = 1, max.pc = 2, scale = TRUE, center = TRUE)
## === Prediction === ##
qd2 <- pcaQDA(formula = Species ~., data = iris, n.pc = 1, max.pc = 2,
             scale = TRUE, center = TRUE)
set.seed(123)
idx <- sample.int(150, 40)
newdata <- iris[idx, 1:4]
newdata.prediction <- predict(qd2, newdata = newdata, type = "all")

## The confusion matrix
x <- data.frame(TRUE.class = iris$Species[idx],
               PRED.class = newdata.prediction$class)
table(x)

#>             PRED.class
#> TRUE.class   setosa versicolor virginica
#>   setosa         12          0         0
#>   versicolor      0         12         2
#>   virginica       0          1        13